[IDLE-111] feat: Mybook 엔티티 필드명 변경 - isPublic, isExchangeable, isShareable, isDeleted --> showable, exchangeable, shareable, deleted - 접두사로 is가 붙을 시, Lombok @Getter와 mapStruct mapping 문제로 인해 is 접두사 제거 - stars 타입 Double 로 수정

+++ backend/book-service/src/main/java/kr/mybrary/bookservice/mybook/presentation/dto/response/MyBookDetailResponse.java
@@ -31,6 +31,6 @@ public class MyBookDetailResponse {
         private List<String> translators;
         private String publisher;
         private String thumbnailUrl;
-        private Integer stars;
+        private Double stars;
     }
 }
